requirements-locked.txt
audio-separator[gpu]==0.35.2
onnxruntime-gpu==1.19.2
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
numpy==1.26.4
librosa==0.10.2.post1
soundfile==0.12.1
resampy==0.4.3
pyloudnorm==0.1.1
tqdm==4.66.4
click==8.1.7
python-dotenv==1.0.1
rich==13.7.1
jinja2==3.1.4
pydub==0.25.1
rvc-python==0.1.5
